refactor: 파일들 한 폴더로 합침
팀 컨벤션에 따라 components, pages/IssueDetail안에 폴더 구조 없앰
CamelCase로 파일명 변경
