+++ client/src/components/LabelInDetail.jsx
