--- client/src/components/dropDown/index.js
@@ -1,1 +1,0 @@
-export { default } from './dropDown.jsx';
