--- client/src/components/sideMenu/assigneePart/index.js
@@ -1,1 +1,0 @@
-export { default } from './assigneePart.jsx';
