--- client/src/components/sideMenu/commonSidePart/index.js
@@ -1,1 +1,0 @@
-export { default } from './commonSidePart.jsx';
