--- client/src/components/sideMenu/labelPart/index.js
@@ -1,1 +1,0 @@
-export { default } from './labelPart.jsx';
