--- client/src/components/sideMenu/milestonePart/index.js
@@ -1,1 +1,0 @@
-export { default } from './milestonePart.jsx';
