--- client/src/components/userDescriptionDetail/index.js
@@ -1,1 +1,0 @@
-export { default } from './descriptionHeader.jsx';
