--- client/src/pages/IssueDetail/mainPart/index.js
@@ -1,1 +1,0 @@
-export { default } from './issueDetailPart.jsx';
